Encode Old Version
* Not Working Alwats

+++ pkg/model/asset.go
@@ -2,6 +2,6 @@ package model
 
 type Asset struct {
 	Uuid         string    `json:"uuid"`
-	LayoutsCount int       `json:"layouts_count"`
+	LayoutsCount int16     `json:"layouts_count"`
 	Layouts      []*Bounds `json:"layouts"`
 }
