+++ pkg/model/bounds.go
@@ -3,5 +3,5 @@ package model
 type Bounds struct {
 	Center   *Vector3 `json:"center"`
 	Extents  *Vector3 `json:"extents"`
-	Rotation int      `json:"rotation"`
+	Rotation int8     `json:"rotation"`
 }
